+++ app/build.gradle.kts
@@ -41,6 +41,7 @@ android {
 
 dependencies {
   // internal
+  implementation(project(":domain:app"))
   implementation(project(":domain:ui"))
   implementation(project(":common:log"))
