+++ build-logic/convention/src/main/kotlin/TemplateFeatureConventionPlugin.kt
@@ -41,6 +41,7 @@ class TemplateFeatureConventionPlugin : TemplateAndroidConventionPlugin() {
         dependencies {
           val implementation by configurations
           // internal dependencies
+          implementation(project(":domain:app"))
           implementation(project(":domain:ui")) // brings in compose
           implementation(project(":common:log"))
         }
